+++ RELEASE_NOTES.md
@@ -46,6 +46,11 @@ There are some breaking changes, but they shouldn't affect you if you haven't be
 
 * `EmojiCategory` now adds accidentally excluded 😵‍💫 emoji.
 
+### 🗑 Deprecations
+
+* A bunch of `SystemKeyboardLayoutProvider` `hasXXXAlphabeticInput` properties have been deprecated.
+
+
 ### 💥 Breaking changes
 
 * `EmojiKeyboardButton` has renamed `configuration` to `style`.